backend-ms1/src/main/java/ma/zyn/app/service/impl/student/instructor/InstructorStudentServiceImpl.java
package ma.zyn.app.service.impl.student.instructor;


import ma.zyn.app.zynerator.exception.EntityNotFoundException;
import ma.zyn.app.bean.core.instructor.Instructor;
import ma.zyn.app.dao.criteria.core.instructor.InstructorCriteria;
import ma.zyn.app.dao.facade.core.instructor.InstructorDao;
import ma.zyn.app.dao.specification.core.instructor.InstructorSpecification;
import ma.zyn.app.service.facade.student.instructor.InstructorStudentService;
import ma.zyn.app.zynerator.service.AbstractServiceImpl;
import static ma.zyn.app.zynerator.util.ListUtil.*;

import org.springframework.stereotype.Service;
import java.util.List;
import java.util.ArrayList;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;

import ma.zyn.app.zynerator.util.RefelexivityUtil;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.transaction.annotation.Propagation;
import org.springframework.transaction.annotation.Transactional;

import ma.zyn.app.service.facade.student.course.CourseStudentService ;
import ma.zyn.app.bean.core.course.Course ;

import java.time.LocalDateTime;
import ma.zyn.app.zynerator.security.service.facade.UserService;
import ma.zyn.app.zynerator.security.service.facade.RoleService;
import ma.zyn.app.zynerator.security.service.facade.RoleUserService;
import ma.zyn.app.zynerator.security.bean.Role;
import ma.zyn.app.zynerator.security.bean.RoleUser;
import ma.zyn.app.zynerator.security.common.AuthoritiesConstants;
import ma.zyn.app.zynerator.security.service.facade.ModelPermissionUserService;
import java.util.Collection;
import java.util.List;
@Service
public class InstructorStudentServiceImpl implements InstructorStudentService {

    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class, readOnly = false)
    public Instructor update(Instructor t) {
        Instructor loadedItem = dao.findById(t.getId()).orElse(null);
        if (loadedItem == null) {
            throw new EntityNotFoundException("errors.notFound", new String[]{Instructor.class.getSimpleName(), t.getId().toString()});
        } else {
            updateWithAssociatedLists(t);
            dao.save(t);
            return loadedItem;
        }
    }

    public Instructor findById(Long id) {
        return dao.findById(id).orElse(null);
    }


    public Instructor findOrSave(Instructor t) {
        if (t != null) {
            Instructor result = findByReferenceEntity(t);
            if (result == null) {
                return dao.save(t);
            } else {
                return result;
            }
        }
        return null;
    }

    public List<Instructor> findAll() {
        return dao.findAll();
    }

    public List<Instructor> findByCriteria(InstructorCriteria criteria) {
        List<Instructor> content = null;
        if (criteria != null) {
            InstructorSpecification mySpecification = constructSpecification(criteria);
            content = dao.findAll(mySpecification);
        } else {
            content = dao.findAll();
        }
        return content;

    }


    private InstructorSpecification constructSpecification(InstructorCriteria criteria) {
        InstructorSpecification mySpecification =  (InstructorSpecification) RefelexivityUtil.constructObjectUsingOneParam(InstructorSpecification.class, criteria);
        return mySpecification;
    }

    public List<Instructor> findPaginatedByCriteria(InstructorCriteria criteria, int page, int pageSize, String order, String sortField) {
        InstructorSpecification mySpecification = constructSpecification(criteria);
        order = (order != null && !order.isEmpty()) ? order : "desc";
        sortField = (sortField != null && !sortField.isEmpty()) ? sortField : "id";
        Pageable pageable = PageRequest.of(page, pageSize, Sort.Direction.fromString(order), sortField);
        return dao.findAll(mySpecification, pageable).getContent();
    }

    public int getDataSize(InstructorCriteria criteria) {
        InstructorSpecification mySpecification = constructSpecification(criteria);
        mySpecification.setDistinct(true);
        return ((Long) dao.count(mySpecification)).intValue();
    }

    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class, readOnly = false)
	public boolean deleteById(Long id) {
        boolean condition = (id != null);
        if (condition) {
            deleteAssociatedLists(id);
            dao.deleteById(id);
        }
        return condition;
    }

    public void deleteAssociatedLists(Long id) {
        courseService.deleteByInstructorId(id);
    }




    @Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class, readOnly = false)
    public List<Instructor> delete(List<Instructor> list) {
		List<Instructor> result = new ArrayList();
        if (list != null) {
            for (Instructor t : list) {
                if(dao.findById(t.getId()).isEmpty()){
					result.add(t);
				}else{
                    dao.deleteById(t.getId());
                }
            }
        }
		return result;
    }


    public Instructor findWithAssociatedLists(Long id){
        Instructor result = dao.findById(id).orElse(null);
        if(result!=null && result.getId() != null) {
            result.setCourses(courseService.findByInstructorId(id));
        }
        return result;
    }

	@Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class, readOnly = false)
    public List<Instructor> update(List<Instructor> ts, boolean createIfNotExist) {
        List<Instructor> result = new ArrayList<>();
        if (ts != null) {
            for (Instructor t : ts) {
                if (t.getId() == null) {
                    dao.save(t);
                } else {
                    Instructor loadedItem = dao.findById(t.getId()).orElse(null);
                    if (isEligibleForCreateOrUpdate(createIfNotExist, t, loadedItem)) {
                        dao.save(t);
                    } else {
                        result.add(t);
                    }
                }
            }
        }
        return result;
    }


    private boolean isEligibleForCreateOrUpdate(boolean createIfNotExist, Instructor t, Instructor loadedItem) {
        boolean eligibleForCreateCrud = t.getId() == null;
        boolean eligibleForCreate = (createIfNotExist && (t.getId() == null || loadedItem == null));
        boolean eligibleForUpdate = (t.getId() != null && loadedItem != null);
        return (eligibleForCreateCrud || eligibleForCreate || eligibleForUpdate);
    }

    public void updateWithAssociatedLists(Instructor instructor){
    if(instructor !=null && instructor.getId() != null){
        List<List<Course>> resultCourses= courseService.getToBeSavedAndToBeDeleted(courseService.findByInstructorId(instructor.getId()),instructor.getCourses());
            courseService.delete(resultCourses.get(1));
        emptyIfNull(resultCourses.get(0)).forEach(e -> e.setInstructor(instructor));
        courseService.update(resultCourses.get(0),true);
        }
    }








    public Instructor findByReferenceEntity(Instructor t){
        return t==null? null : dao.findByEmail(t.getEmail());
    }



    public List<Instructor> findAllOptimized() {
        return dao.findAllOptimized();
    }

    @Override
    public List<List<Instructor>> getToBeSavedAndToBeDeleted(List<Instructor> oldList, List<Instructor> newList) {
        List<List<Instructor>> result = new ArrayList<>();
        List<Instructor> resultDelete = new ArrayList<>();
        List<Instructor> resultUpdateOrSave = new ArrayList<>();
        if (isEmpty(oldList) && isNotEmpty(newList)) {
            resultUpdateOrSave.addAll(newList);
        } else if (isEmpty(newList) && isNotEmpty(oldList)) {
            resultDelete.addAll(oldList);
        } else if (isNotEmpty(newList) && isNotEmpty(oldList)) {
			extractToBeSaveOrDelete(oldList, newList, resultUpdateOrSave, resultDelete);
        }
        result.add(resultUpdateOrSave);
        result.add(resultDelete);
        return result;
    }

    private void extractToBeSaveOrDelete(List<Instructor> oldList, List<Instructor> newList, List<Instructor> resultUpdateOrSave, List<Instructor> resultDelete) {
		for (int i = 0; i < oldList.size(); i++) {
                Instructor myOld = oldList.get(i);
                Instructor t = newList.stream().filter(e -> myOld.equals(e)).findFirst().orElse(null);
                if (t != null) {
                    resultUpdateOrSave.add(t); // update
                } else {
                    resultDelete.add(myOld);
                }
            }
            for (int i = 0; i < newList.size(); i++) {
                Instructor myNew = newList.get(i);
                Instructor t = oldList.stream().filter(e -> myNew.equals(e)).findFirst().orElse(null);
                if (t == null) {
                    resultUpdateOrSave.add(myNew); // create
                }
            }
	}

    @Override
    public String uploadFile(String checksumOld, String tempUpladedFile, String destinationFilePath) throws Exception {
        return null;
    }


    @Override
    public Instructor create(Instructor t) {
        if (findByUsername(t.getUsername()) != null || t.getPassword() == null) return null;
        t.setPassword(userService.cryptPassword(t.getPassword()));
        t.setEnabled(true);
        t.setAccountNonExpired(true);
        t.setAccountNonLocked(true);
        t.setCredentialsNonExpired(true);
        t.setPasswordChanged(false);

        Role role = new Role();
        role.setAuthority(AuthoritiesConstants.INSTRUCTOR);
        role.setCreatedAt(LocalDateTime.now());
        Role myRole = roleService.create(role);
        RoleUser roleUser = new RoleUser();
        roleUser.setRole(myRole);
        if (t.getRoleUsers() == null)
            t.setRoleUsers(new ArrayList<>());

        t.getRoleUsers().add(roleUser);
        if (t.getModelPermissionUsers() == null)
            t.setModelPermissionUsers(new ArrayList<>());

        t.setModelPermissionUsers(modelPermissionUserService.initModelPermissionUser());

        Instructor mySaved = dao.save(t);

        if (t.getModelPermissionUsers() != null) {
            t.getModelPermissionUsers().forEach(e -> {
                e.setUser(mySaved);
                modelPermissionUserService.create(e);
            });
        }
        if (t.getRoleUsers() != null) {
            t.getRoleUsers().forEach(element-> {
                element.setUser(mySaved);
                roleUserService.create(element);
            });
        }

        if (t.getCourses() != null) {
        t.getCourses().forEach(element-> {
            element.setInstructor(mySaved);
            courseService.create(element);
        });
        }
        return mySaved;
     }

    public Instructor findByUsername(String username){
        return dao.findByUsername(username);
    }

    public boolean changePassword(String username, String newPassword) {
        return userService.changePassword(username, newPassword);
    }




    private @Autowired UserService userService;
    private @Autowired RoleService roleService;
    private @Autowired ModelPermissionUserService modelPermissionUserService;
    private @Autowired RoleUserService roleUserService;

    @Autowired
    private CourseStudentService courseService ;

    public InstructorStudentServiceImpl(InstructorDao dao) {
        this.dao = dao;
    }

    private InstructorDao dao;
}
